refactor permutation mutation operator tests
